# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "CommentCreate": {
        "properties": {
          "content": {
            "maxLength": 2000,
            "title": "Content",
            "type": "string"
          }
        },
        "required": [
          "content"
        ],
        "title": "CommentCreate",
        "type": "object"
      },
      "CommentResponse": {
        "properties": {
          "ai_has_concern": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Has Concern"
          },
          "content": {
            "maxLength": 2000,
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_ai_generated": {
            "default": false,
            "title": "Is Ai Generated",
            "type": "boolean"
          },
          "user_display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Display Name"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "user_role": {
            "title": "User Role",
            "type": "string"
          }
        },
        "required": [
          "content",
          "id",
          "user_id",
          "user_display_name",
          "user_role",
          "created_at"
        ],
        "title": "CommentResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MilestoneCreate": {
        "properties": {
          "achieved_date": {
            "format": "date",
            "title": "Achieved Date",
            "type": "string"
          },
          "image_urls": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Image Urls",
            "type": "array"
          },
          "milestone_type": {
            "maxLength": 50,
            "title": "Milestone Type",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "title": {
            "maxLength": 100,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "milestone_type",
          "title",
          "achieved_date"
        ],
        "title": "MilestoneCreate",
        "type": "object"
      },
      "MilestoneResponse": {
        "properties": {
          "achieved_date": {
            "format": "date",
            "title": "Achieved Date",
            "type": "string"
          },
          "baby_id": {
            "title": "Baby Id",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "image_urls": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Image Urls",
            "type": "array"
          },
          "milestone_type": {
            "maxLength": 50,
            "title": "Milestone Type",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "title": {
            "maxLength": 100,
            "title": "Title",
            "type": "string"
          },
          "unlocked_achievements": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/UnlockedAchievementInfo"
            },
            "title": "Unlocked Achievements",
            "type": "array"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "required": [
          "milestone_type",
          "title",
          "achieved_date",
          "id",
          "baby_id",
          "user_id"
        ],
        "title": "MilestoneResponse",
        "type": "object"
      },
      "MilestoneTimelineGroup": {
        "properties": {
          "milestones": {
            "items": {
              "$ref": "#/components/schemas/MilestoneResponse"
            },
            "title": "Milestones",
            "type": "array"
          },
          "month_age": {
            "title": "Month Age",
            "type": "integer"
          }
        },
        "required": [
          "month_age",
          "milestones"
        ],
        "title": "MilestoneTimelineGroup",
        "type": "object"
      },
      "MilestoneUpdate": {
        "properties": {
          "achieved_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Achieved Date"
          },
          "image_urls": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Urls"
          },
          "milestone_type": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Milestone Type"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "MilestoneUpdate",
        "type": "object"
      },
...
      "UnlockedAchievementInfo": {
        "description": "記録APIレスポンスに含める実績解除情報",
        "properties": {
          "achievement_id": {
            "title": "Achievement Id",
            "type": "string"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "icon": {
            "title": "Icon",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "rarity": {
            "title": "Rarity",
            "type": "string"
          }
        },
        "required": [
          "achievement_id",
          "name",
          "icon",
          "category",
          "rarity",
          "description"
        ],
        "title": "UnlockedAchievementInfo",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Botoro API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/comments/{comment_id}": {
      "delete": {
        "operationId": "delete_comment_api_comments__comment_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "comment_id",
            "required": true,
            "schema": {
              "title": "Comment Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Comment",
        "tags": [
          "comments"
        ]
      }
    },
...
    "/api/milestones/": {
      "get": {
        "operationId": "get_milestones_api_milestones__get",
        "parameters": [
          {
            "in": "query",
            "name": "baby_id",
            "required": true,
            "schema": {
              "title": "Baby Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MilestoneResponse"
                  },
                  "title": "Response Get Milestones Api Milestones  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Milestones",
        "tags": [
          "milestones"
        ]
      },
      "post": {
        "operationId": "create_milestone_api_milestones__post",
        "parameters": [
          {
            "in": "query",
            "name": "baby_id",
            "required": true,
            "schema": {
              "title": "Baby Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MilestoneCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MilestoneResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Milestone",
        "tags": [
          "milestones"
        ]
      }
    },
    "/api/milestones/timeline": {
      "get": {
        "operationId": "get_milestone_timeline_api_milestones_timeline_get",
        "parameters": [
          {
            "in": "query",
            "name": "baby_id",
            "required": true,
            "schema": {
              "title": "Baby Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MilestoneTimelineGroup"
                  },
                  "title": "Response Get Milestone Timeline Api Milestones Timeline Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Milestone Timeline",
        "tags": [
          "milestones"
        ]
      }
    },
    "/api/milestones/{milestone_id}": {
      "delete": {
        "operationId": "delete_milestone_api_milestones__milestone_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "milestone_id",
            "required": true,
            "schema": {
              "title": "Milestone Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Milestone",
        "tags": [
          "milestones"
        ]
      },
      "patch": {
        "operationId": "update_milestone_api_milestones__milestone_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "milestone_id",
            "required": true,
            "schema": {
              "title": "Milestone Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MilestoneUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MilestoneResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Milestone",
        "tags": [
          "milestones"
        ]
      }
    },
...
    "/api/records/{record_type}/{record_id}/comments": {
      "get": {
        "operationId": "get_record_comments_api_records__record_type___record_id__comments_get",
        "parameters": [
          {
            "in": "path",
            "name": "record_type",
            "required": true,
            "schema": {
              "title": "Record Type",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "record_id",
            "required": true,
            "schema": {
              "title": "Record Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CommentResponse"
                  },
                  "title": "Response Get Record Comments Api Records  Record Type   Record Id  Comments Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Record Comments",
        "tags": [
          "comments"
        ]
      },
      "post": {
        "operationId": "create_record_comment_api_records__record_type___record_id__comments_post",
        "parameters": [
          {
            "in": "path",
            "name": "record_type",
            "required": true,
            "schema": {
              "title": "Record Type",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "record_id",
            "required": true,
            "schema": {
              "title": "Record Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Record Comment",
        "tags": [
          "comments"
        ]
      }
    },
...
  }
}